requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
autopep8==1.7.0
backports.asyncio.runner==1.2.0
bitarray==3.7.2
cachetools==6.2.1
cairo-lang==0.14.0.1
certifi==2025.10.5
cffi==1.15.1
charset-normalizer==3.4.4
ckzg==2.1.5
coverage==7.10.6
cryptography==38.0.4
cytoolz==1.1.0
dataclasses-json==0.5.7
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.0
execnet==2.1.1
fastecdsa==3.0.1
frozendict==2.4.6
frozenlist==1.8.0
future==1.0.0
gprof2dot==2025.4.14
hexbytes==1.3.1
idna==3.11
iniconfig==2.1.0
lark==1.3.1
liboqs-python @ file:///home/upasana/221IT075/Major%20Project/MajorProject_endsem_v1/liboqs-python
marshmallow==3.17.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.2.0
marshmallow_dataclass==8.6.1
mpmath==1.3.0
msgpack-python==0.5.6
multidict==6.7.0
mypy-extensions==0.4.3
numpy==2.2.6
oqs==0.10.2
packaging==25.0
parsimonious==0.10.0
petlib==0.0.45
pipdeptree==2.29.0
pluggy==1.6.0
prometheus_client==0.23.1
propcache==0.4.1
pyasn1==0.6.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.16.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.0.9
pyring==0.0.2
pytest==8.4.1
pytest-asyncio==1.2.0
pytest-cov==6.2.1
pytest-profiling==1.8.1
pytest-xdist==3.8.0
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rlp==4.1.0
six==1.17.0
sympy==1.14.0
toml==0.10.2
tomli==2.2.1
toolz==1.1.0
typeguard==2.13.3
types-requests==2.32.4.20250913
typing-inspect==0.8.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
